Split video into SDL 1.2 to 2
