+++ client/sdl/i_input.cpp
@@ -27,6 +27,7 @@
 #include <stdlib.h>
 #include <list>
 
+#include "i_sdl.h"
 #include "win32inc.h"
 
 #include "doomstat.h"
@@ -44,6 +45,12 @@
 	#include "i_xbox.h"
 #elif __SWITCH__
 	#include "nx_io.h"
+#endif
+
+#if defined(SDL12)
+#include "i_input_sdl12.h"
+#elif defined(SDL20)
+#include "i_input_sdl20.h"
 #endif
 
 #ifdef _WIN32
